+++ src/cone/ugm/browser/groups.py
@@ -6,7 +6,6 @@ from cone.ugm.browser.columns import Column
 from cone.ugm.browser.listing import PrincipalsListing
 from cone.ugm.model.groups import Groups
 from pyramid.i18n import TranslationStringFactory
-from pyramid.security import has_permission
 import logging
 
 
@@ -14,8 +13,11 @@ logger = logging.getLogger('cone.ugm')
 _ = TranslationStringFactory('cone.ugm')
 
 
-@tile('leftcolumn', 'templates/principals_left_column.pt',
-      interface=Groups, permission='view')
+@tile(
+    name='leftcolumn',
+    path='templates/principals_left_column.pt',
+    interface=Groups,
+    permission='view')
 class GroupsLeftColumn(Tile):
     add_label = _('add_group', default='Add Group')
 
@@ -29,11 +31,14 @@ class GroupsLeftColumn(Tile):
 
     @property
     def can_add(self):
-        return has_permission('add_group', self.model, self.request)
+        return self.request.has_permission('add_group', self.model)
 
 
-@tile('rightcolumn', 'templates/principals_right_column.pt',
-      interface=Groups, permission='view')
+@tile(
+    name='rightcolumn',
+    path='templates/principals_right_column.pt',
+    interface=Groups,
+    permission='view')
 class GroupsRightColumn(Column):
 
     @property
@@ -49,13 +54,16 @@ class GroupsRightColumn(Column):
         return make_url(self.request, node=self.model[self.principal_id])
 
 
-@tile('columnlisting', 'templates/column_listing.pt',
-      interface=Groups, permission='view')
+@tile(
+    name='columnlisting',
+    path='templates/column_listing.pt',
+    interface=Groups,
+    permission='view')
 class GroupsColumnListing(PrincipalsListing):
     slot = 'leftlisting'
     list_columns = PrincipalsListing.group_list_columns
     listing_attrs = PrincipalsListing.group_attrs
-    listing_criteria = PrincipalsListing.group_listing_criteria
+    localmanager_ids = PrincipalsListing.group_localmanager_ids
     sort_attr = PrincipalsListing.group_default_sort_column
     css = 'groups'
     batchname = 'leftbatch'