+++ src/cone/ugm/browser/portrait.py
@@ -7,7 +7,6 @@ from plumber import default
 from plumber import plumb
 from pyramid.i18n import TranslationStringFactory
 from pyramid.response import Response
-from pyramid.security import has_permission
 from pyramid.view import view_config
 from yafowil.base import UNSET
 from yafowil.base import factory
@@ -16,7 +15,10 @@ from yafowil.base import factory
 _ = TranslationStringFactory('cone.ugm')
 
 
-@view_config('portrait_image', context=User, permission='view_portrait')
+@view_config(
+    name='portrait_image',
+    context=User,
+    permission='view_portrait')
 def portrait_image(model, request):
     """XXX: needs polishing. Return configured default portrait if not set
     on user.
@@ -49,7 +51,7 @@ class PortraitForm(Behavior):
             return
         model = self.model
         request = self.request
-        if has_permission('edit_user', model.parent, request):
+        if request.has_permission('edit_user', model.parent):
             mode = 'edit'
         else:
             mode = 'display'
@@ -92,7 +94,7 @@ class PortraitForm(Behavior):
     @plumb
     def save(_next, self, widget, data):
         if not self.portrait_support or \
-          not has_permission('edit_user', self.model.parent, self.request):
+                not self.request.has_permission('edit_user', self.model.parent):
             _next(self, widget, data)
             return
         cfg = ugm_general(self.model)